% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{dah}
\alias{dah}
\title{Dummy function in order to have import to namespace done.}
\usage{
dah(dummy)
}
\arguments{
\item{dummy}{A dummy variable.}
}
\description{
The comments of the function get imported or the operator aren't imported
otherwise.
}
