% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{data4plot}
\alias{data4plot}
\title{Prepare data for plot lollipop, ts, mosaic, RVP.}
\usage{
data4plot(data, keepSize, keepSource, RRL, RRV, RSL, SPL, SPV, SSL,
  yearRef, rankBy, topN, ditch, showFY, lollipop_measure, ts_measure,
  mosaic_measure, RVP_facetY)
}
\arguments{
\item{data}{a dataframe.}

\item{keepSize}{a boolean.}

\item{keepSource}{a boolean.}

\item{RRL}{a string, reportRegionLevel.}

\item{RRV}{a string, ReportRegionValue.}

\item{RSL}{a string, RegionSplitLevel.}

\item{SPL}{a string, SuppliesPlotLevel.}

\item{SPV}{a string, SuppliesPlotValue.}

\item{SSL}{a string, SppliesSplitLevel.}

\item{yearRef}{an Integer representing the reference year.}

\item{rankBy}{a string representing the measure on which the Region are
ranked, "Rev" or "Vol".}

\item{topN}{An integer representing the number of items to display, the
others will be grouped together.}

\item{ditch}{a string, a column name.}

\item{showFY}{a string vector, such as "All" or c("2017", "2018").}

\item{lollipop_measure}{a string, such as "Rev" or "Liters". NULL if the
plot isn't a lollipop.}

\item{ts_measure}{a string, such as "Rev_sm" or "Liters_sm". NULL if not a
ts plot.}

\item{mosaic_measure}{a string, such as "Rev" or "Liters". NULL if the
plot isn't a mosaic.}

\item{RVP_facetY}{a string, such as "RegionSplit", "source". NULL is the
plot isn't an RVP.}
}
\value{
a dataframe expected by one of the plot functions.
}
\description{
\code{data4plot} regroups several functions, in order:
\code{\link{f_keepSize}}, \code{\link{f_keepSource}},
\code{\link{f_keepRegion}}, \code{\link{f_keepSupplies}},
\code{\link{f_rkRegion}}, \code{\link{f_rkSupplies}},
\code{\link{f_topSupplies}}, \code{\link{f_keepDateYear}},
\code{\link{f_summariseIFnum}}, \code{\link{f_summarise4lollipop}},
\code{\link{f_summarise4ts}}, \code{\link{f_summarise4mosaic}},
\code{\link{f_summarise4RVP}}.
\code{data4plot} is used in:
\code{\link{data4plot_RVP}}
\code{\link{data4plot_Lollipop}}
\code{\link{data4plot_Mosaic}}
\code{\link{data4plot_ts}}
}
