% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{data4plot_Lollipop}
\alias{data4plot_Lollipop}
\title{Returns data for plot_lollipop.}
\usage{
data4plot_Lollipop(data, RRL, RRV, RSL, SPL, SPV, SSL, yearRef, rankBy,
  topN, showFY, lollipop_measure, SP_filter_part)
}
\arguments{
\item{SP_filter_part}{a string helping to selct a regular expression in the
dataframe data_SL2_regex.}
}
\value{
a list containing a dataframe, data, and a string vector, labs.
}
\description{
\code{data4plot_Lollipop} uses \code{\link{data4plot}}.
}
