% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{data4plot_RVP}
\alias{data4plot_RVP}
\title{Returns data for plot_RVP.}
\usage{
data4plot_RVP(data, keepSize, keepSource, RRL, RRV, RSL, SPL, SPV, SSL,
  yearRef, rankBy, topN, RVP_facetY, SP_filter_part, free_y)
}
\arguments{
\item{SP_filter_part}{a string helping to selct a regular expression in the
dataframe data_SL2_regex.}

\item{free_y}{a string, such as "free_y".}
}
\value{
a list containing a dataframe, data, and a string vector, labs.
}
\description{
\code{data4plot_RVP} uses \code{\link{data4plot}}.
}
