% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{data4plot_ts}
\alias{data4plot_ts}
\title{Returns data for plot_ts.}
\usage{
data4plot_ts(data, RRL, RRV, RSL, SPL, SPV, SSL, yearRef, rankBy, topN,
  ts_measure, SP_filter_part, free_y)
}
\arguments{
\item{SP_filter_part}{a string helping to selct a regular expression in the
dataframe data_SL2_regex.}

\item{free_y}{a string, such as "free_y".}
}
\value{
a list containing a dataframe, data, and a string vector, labs.
}
\description{
\code{data4plot_ts} uses \code{\link{data4plot}}.
}
