% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_attrition.R
\docType{data}
\name{data_attrition}
\alias{data_attrition}
\title{attrition data to plot.}
\format{a data frame
\describe{
\item{SuppliesLevel4}{data frame fit rate}
\item{attritionProfile}{name of the attrition profile}
\item{model}{models of hw}
\item{seg}{parameters for 3 segments of the fleets}
}}
\source{
computation from inkSales and the hierarchy of region and product
}
\usage{
data_attrition
}
\description{
SuppliesLevel4, model, attrition profile and parameters.
}
\keyword{datasets}
