% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_attrition.R
\name{data_plot_attrition}
\alias{data_plot_attrition}
\title{Return a dataframe, with 1 ggplot2 object in column plot.}
\usage{
data_plot_attrition(SuppliesLevel4_id, data_attrition)
}
\arguments{
\item{SuppliesLevel4_id}{a string, select the model to plot.}

\item{data_attrition}{a dataframe with columns: SuppliesLevel4, hwRatio,
attritionProfile, hwmodel2, segX_percent, segX_begin, segX_end.}
}
\value{
a dataframe with columns, SuppliesLevel4, plot.
}
\description{
\code{data_plot_attrition} uses \code{\link{plot_3segAttrition}}.
}
