% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sku.R
\docType{data}
\name{data_sku}
\alias{data_sku}
\title{list of skus and matching SuppliesLevel4.}
\format{a data frame
\describe{
\item{SuppliesLevel4}{aggregation level for fit rate calculation}
\item{Material}{sku}
\item{SAP_Model}{aggregation in SAP}
}}
\source{
source data for the agregation process of the products
}
\usage{
data_sku
}
\description{
based on RVP data.
}
\keyword{datasets}
