% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_treemap.R
\name{dataTree}
\alias{dataTree}
\title{Prepare data for plotting treemap.}
\usage{
dataTree(dat, cat, subcat, RevVol, FY)
}
\arguments{
\item{dat}{A dataframe with columns evaluate(RevVol), fDate, evaluate(cat),
evaluate(subcat).}

\item{cat}{A string.}

\item{subcat}{A string.}

\item{RevVol}{A string.}

\item{FY}{An integer, fiscal year.}
}
\value{
a dataframe ready for plotting by plotTree(), columns category,
  subcategory and total.
}
\description{
\code{dataTree}
}
