% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{divid}
\alias{divid}
\title{Returns the order or magnitude of a double (1, 1 000, 100 000, ...).}
\usage{
divid(y)
}
\arguments{
\item{y}{a double.}
}
\value{
an integer representing the order of magnitude.
}
\description{
\code{divid} is used \code{\link{data4plot_Lollipop}}.
}
