% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_attrition.R
\name{f_basicAttrition}
\alias{f_basicAttrition}
\title{Returns the (1-attrition) profile, month by month.}
\usage{
f_basicAttrition(month_at_100 = 24, month_0_reached = 48)
}
\arguments{
\item{month_at_100}{an integer, month when decline starts.}

\item{month_0_reached}{an integer, month when decline ends.}
}
\value{
a double vector, with continuous declining values from 1 to 0.
}
\description{
\code{f_basicAttrition}
}
