% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{f_C2Fdate}
\alias{f_C2Fdate}
\title{Convert calendar date to fiscal date.}
\usage{
f_C2Fdate(cDate, table = date_convert)
}
\arguments{
\item{cDate}{a calendar date.}

\item{table}{A tibble representing months.}
}
\value{
calendar date
}
\description{
\code{f_C2Fdate} uses date convert dataframe (in global environment) to
convert the calendar date input in fiscal date output.
}
