% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{f_df2v}
\alias{f_df2v}
\title{Transform a 2-columns df into a vector with named items.}
\usage{
f_df2v(df)
}
\arguments{
\item{df}{a 2-columns dataframe. Column1 is the name, column2 the value.}
}
\value{
a vecto of named items.
}
\description{
\code{f_df2v} transform a 2-columns dataframe into a vector.
Used for instance with date convertions (calendat to/from fiscal).
}
