% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{f_F2Cdate}
\alias{f_F2Cdate}
\title{Convert fiscal date to calendar date.}
\usage{
f_F2Cdate(fDate, table = date_convert)
}
\arguments{
\item{fDate}{a fiscal date.}

\item{table}{A tibble representing months.}
}
\value{
calendar date
}
\description{
\code{f_F2Cdate} uses date convert dataframe (in global environment) to
convert the fiscal date input in calendar date output.
}
