% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_fr.R
\name{f_fr_merge}
\alias{f_fr_merge}
\title{Join Inksales and MIF data, with prior relevant summary.}
\usage{
f_fr_merge(IS, MIF)
}
\arguments{
\item{IS}{A dataframe, ink sales}

\item{MIF}{A dataframe, Machine In Field}
}
\value{
a dataframe with columns: RL0, 2, 4, fDate, SL4, cumqty, cumMIF,
  Liters_sm
}
\description{
\code{f_fr_merge} used only in Supplies Project not in suppliesdash.
}
