% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{f_keepDateYear}
\alias{f_keepDateYear}
\title{Removes a column to a dataframe.}
\usage{
f_keepDateYear(data, ditch = "fYear")
}
\arguments{
\item{data}{a dataframe}

\item{ditch}{a string, a column name.}
}
\value{
a dataframe without the column ditch.
}
\description{
\code{f_keepDateYear} removes a column to a dataframe.
Used in \code{\link{data4plot}}.
}
