% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{f_keepRegion}
\alias{f_keepRegion}
\title{Keep the relevant region columns (1 or 2 distinct).}
\usage{
f_keepRegion(data, ReportRegionLevel, ReportRegionValue, RegionSplitLevel)
}
\arguments{
\item{data}{A dataframe with columns RegionLevel0, 1, 2, 4.}

\item{ReportRegionLevel}{A string, such as "RegionLevel0".}

\item{ReportRegionValue}{A string, such as "EMEAR".}

\item{RegionSplitLevel}{A string, such as "RegionLevel2".}
}
\value{
a dataframe with columns RegionReport, RegionSplit.
}
\description{
\code{f_keepRegion} is used in \code{\link{data4plot}}.
If ReportRegionLevel == RegionSplitLevel, the 2 output region colums will
have the same value.
NA values and "PEA" are removed silently.
}
