% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{f_keepSize}
\alias{f_keepSize}
\title{Remove, or not, the size column.}
\usage{
f_keepSize(data, keepSize)
}
\arguments{
\item{data}{A dataframe}

\item{keepSize}{A boolean}
}
\value{
a dataframe.
}
\description{
\code{f_keepSize} is used in \code{\link{data4plot}}.
}
