% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{f_keepSource}
\alias{f_keepSource}
\title{Remove, or not, the column source.}
\usage{
f_keepSource(data, keepSource)
}
\arguments{
\item{data}{A dataframe}

\item{keepSource}{A boolean}
}
\value{
a dataframe.
}
\description{
\code{f_keepSource} is used in \code{\link{data4plot}}.
In \code{f_keepSource}, if the source column is kept, it is mutated into an
ordered factor.
}
