% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{f_keepSupplies}
\alias{f_keepSupplies}
\title{Keep the relevant supplies columns (1 or 2 distinct).}
\usage{
f_keepSupplies(data, SuppliesPlotLevel, SuppliesPlotValue,
  SuppliesSplitLevel)
}
\arguments{
\item{data}{a dataframe with columns SuppliesLevel0, 1, 2, 3, 4.}

\item{SuppliesPlotLevel}{a string, such as "SuppliesLevel0".}

\item{SuppliesPlotValue}{a string, such as "LFP aqueous ink".}

\item{SuppliesSplitLevel}{a string, such as "SuppliesLevel3".}
}
\value{
a dataframe with columns SuppliesPlot, SuppliesSplit.
}
\description{
\code{f_keepSupplies} is used in \code{\link{data4plot}}.
If SuppliesPlotLevel == SuppliesSplitLevel, the 2 output supplies columns
are identical.
NA values are removed silently.
}
