% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{f_rkRegion}
\alias{f_rkRegion}
\title{Rank Region according to revenue (or volume) within the reference year.}
\usage{
f_rkRegion(data, yearRef = 2017, rankBy = "Rev")
}
\arguments{
\item{data}{a dataframe with columns: fDate, RegionSplit, Rev, Vol.}

\item{yearRef}{an Integer representing the reference year.}

\item{rankBy}{a string representing the measure on which the Region are
ranked, "Rev" or "Vol".}
}
\value{
a dataframe with an ordered factor for column RegionSplit and
  column fYear.
}
\description{
\code{f_rkRegion} is used in \code{\link{data4plot}}.
}
