% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{f_summarise4lollipop}
\alias{f_summarise4lollipop}
\title{Prepare data for lollipop plots.}
\usage{
f_summarise4lollipop(data, lollipop_measure = "Rev", showFY)
}
\arguments{
\item{data}{a dataframe with column fYear, lollipop_measure, regionSplit,
SuppliesSplit}

\item{lollipop_measure}{a string, such as "Rev" or "Liters". NULL if the
plot isn't a lollipop.}

\item{showFY}{a string vector, such as "All" or c("2017", "2018").}
}
\value{
a dataframe with columns RegionSplit, SuppliesSplit, fYear,
  percentage, measure.
}
\description{
\code{f_summarise4lollipop} adds information at level (RegionSplit, fYear),
and level (RegionSplit, SuppliesSplit, fYear), while keeping all the rows of
the data frame.
Used in \code{\link{data4plot}}.
}
