% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{f_summarise4RVP}
\alias{f_summarise4RVP}
\title{Prepare data for RVP plots.}
\usage{
f_summarise4RVP(data, facetY)
}
\arguments{
\item{data}{a dataframe with columns SuppliesSplit, RegionReport, Supplies,
facetY, Rev, Liters, Rev_sm, Liters_sm, Date, fDate.}

\item{facetY}{a string, such as "RegionSplit", "source". NULL is the plot
isn't an RVP.}
}
\value{
a dataframe with columns SuppliesSplit, RegionReport, Supplies,
  facetY, KEY, VALUE, Date, fDate, label. KEY (Rev_sm, Liters_sm, price) a
  factor.
}
\description{
\code{f_summarise4RVP} adds information at level (RegionSplit, fYear),
and level (RegionSplit, SuppliesSplit, fYear), while keeping all the rows of
the data frame.
Requires to drop fYear before (\code{f_keepDateYear}).
}
