% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{f_summarise4ts}
\alias{f_summarise4ts}
\title{Prepare data for time series plots.}
\usage{
f_summarise4ts(data, ts_measure = "Rev_sm")
}
\arguments{
\item{data}{a dataframe with columns fDate, ts_measure, SuppliesSplit.}

\item{ts_measure}{a string, such as "Rev_sm" or "Liters_sm". NULL if not a
ts plot.}
}
\value{
a dataframe with column fDate, measure, SuppliesSplit.
}
\description{
\code{f_summarise4ts} filters out old data ans a few old models.
Identify the column measure.
Used in \code{\link{data4plot}}.
}
