% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{f_summariseIFnum}
\alias{f_summariseIFnum}
\title{Summarise all numeric columns (sum)}
\usage{
f_summariseIFnum(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a dataframe
}
\description{
\code{f_summariseIFnum} is grouping by the non numeric columns and
summarises, with a sum, the numeric columns. If there is a column fYear
the column is converted in text during the calculation and return as an
integer.
Used in \code{\link{data4plot}}.
}
