% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_RVP.R
\name{f_topSupplies}
\alias{f_topSupplies}
\title{Keep the topN supplies items and lump together the Others.}
\usage{
f_topSupplies(data, topN = 10)
}
\arguments{
\item{data}{A dataframe with columns rk, SuppliesSplit.}

\item{topN}{An integer representing the number of items to display, the
others will be grouped together.}
}
\value{
a dataframe with a column SuppliesSplit, an ordered factor.
}
\description{
\code{f_topSupplies} keeps the topN items on SuppliesSplit column.
Used in \code{\link{data4plot}}.
}
