% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_waterfall.R
\name{f_waterfall_input}
\alias{f_waterfall_input}
\title{Prepare the data for the waterfal chart.}
\usage{
f_waterfall_input(data, Driver, cal_period, topN_driver = NULL)
}
\arguments{
\item{data}{A dataframe.}

\item{Driver}{A string.}

\item{cal_period}{A date range.}

\item{topN_driver}{Integer, max number of positive drivers positive, and
max number of negative drivers.}
}
\value{
a dataframe with columns: driver, DeltaValue, balance, time, flow.
}
\description{
\code{f_waterfall_input} use \code{\link{label_FYdate}} and
\code{\link{order_driver}}.
}
