% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{finer}
\alias{finer}
\title{Return the following finer level.}
\usage{
finer(x, vx)
}
\arguments{
\item{x}{A string, an item from vector Vx.}

\item{vx}{A vector.}
}
\value{
A string.
}
\description{
\code{finer} is used in the treemap.
}
