% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{flatten_in_order}
\alias{flatten_in_order}
\title{Convert a dataframe representing a hierarchy into a vector.}
\usage{
flatten_in_order(df)
}
\arguments{
\item{df}{a dataframe representing a hierarchy.}
}
\value{
a vector.
}
\description{
\code{flatten_in_order} is recursive.
\code{flatten_in_order} is used in \code{\link{set_names_vect}}.
}
