% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{fmt_decimals}
\alias{fmt_decimals}
\title{Format continuous axis in ggplot2.}
\usage{
fmt_decimals(decimals = 0)
}
\arguments{
\item{decimals}{number of decimals to keep}
}
\value{
a formated numeric
}
\description{
\code{fmt_decimals} is rounding and forcing non-scientific numeric format.
}
