% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_waterfall.R
\name{label_FYdate}
\alias{label_FYdate}
\title{Make a name (string) for a time interval.}
\usage{
label_FYdate(x)
}
\arguments{
\item{x}{a time interval.}
}
\value{
a string representing the time interval in input.
}
\description{
\code{label_FYdate} is used in \code{\link{f_waterfall_input}}.
}
