% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launches the suppliesdash app.}
\usage{
launchApp()
}
\value{
shiny application object
}
\description{
Display a shiny user interface that display graphs and filters.
}
