% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{leadingDots}
\alias{leadingDots}
\title{Add leading dots to the items of an ordered hierarchical dataframe.}
\usage{
leadingDots(df)
}
\arguments{
\item{df}{a dataframe representing a hierarchy.}
}
\value{
a dataframe.
}
\description{
\code{leadingDots} is used in \code{\link{set_names_vect}}.
Uses function \code{\link{dotsVector}}.
}
