% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{multinest}
\alias{multinest}
\title{function found on the web.}
\usage{
multinest(t, no_of_name_cols)
}
\arguments{
\item{t}{what is it already}

\item{no_of_name_cols}{an integer.
return a list.}
}
\description{
\code{multinest} recursive function. not used.
}
