% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_waterfall.R
\name{order_driver}
\alias{order_driver}
\title{Prepare the data for f_waterfall_input depending on the Driver value.}
\usage{
order_driver(df, Driver, max_driver_count)
}
\arguments{
\item{df}{dataframe, column are driver and DeltaValue.}

\item{Driver}{A string.}

\item{max_driver_count}{Integer, same as topN_driver.}
}
\value{
a dataframe with the same column as the input.
}
\description{
\code{order_driver} is used in \code{\link{f_waterfall_input}}.
No modification are done to the data in the case the driver if fDate.
}
