% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_attrition.R
\name{plot_3segAttrition}
\alias{plot_3segAttrition}
\title{Return a ggplot2 object displaying the 3 segment attrition profile.}
\usage{
plot_3segAttrition(SuppliesLevel4, attritionProfile, hwmodel2,
  segA_percent = c(0.1, 0.3), segA_begin = 24, segA_end = 48,
  segB_percent = c(0.7, 0.4), segB_begin = 36, segB_end = 120,
  segC_percent = c(0.2, 0.3), segC_begin = 60, segC_end = 240)
}
\arguments{
\item{SuppliesLevel4}{a string}

\item{attritionProfile}{astrin TBC}

\item{hwmodel2}{a string TBC}

\item{segA_percent}{a double, the share of the population in segment A.}

\item{segA_begin}{an integer, month where decline starts for segment A.}

\item{segA_end}{an integer, month where decline ends for segment A.}

\item{segB_percent}{a double, the share of the population in segment B.}

\item{segB_begin}{an integer, month where decline starts for segment B.}

\item{segB_end}{an integer, month where decline ends for segment B.}

\item{segC_percent}{a double, the share of the population in segment C.}

\item{segC_begin}{an integer, month where decline starts for segment C.}

\item{segC_end}{an integer, month where decline ends for segment C.}
}
\value{
a ggplot2 object.
}
\description{
\code{plot_3segAttrition} uses \code{\link{f_basicAttrition}}.
grey dashed lines shows the half-life.
}
