% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fr.R
\name{plot_fr}
\alias{plot_fr}
\title{Build a ggplot2 object with geom_line and facet on KEY.}
\usage{
plot_fr(FR)
}
\arguments{
\item{FR}{a dataframe with columns: fDate, VALUE, regionSplit, KEY, supplies.}
}
\value{
a ggplot2 object.
}
\description{
\code{plot_fr} expects input data from \code{\link{f_fr_filter}}.
One unique supplies (SuppliesLevel4) is selected for this plot.
}
