% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RVP.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Returns a ggplot2 object.}
\usage{
plot_lollipop(dat)
}
\arguments{
\item{dat}{A list including a dataframe, data, and labels, labs.
data has columns: percentage, SuppliesSplit, measure, RegionSplit, fYear.}
}
\value{
a ggplot2 object.
}
\description{
\code{plot_lollipop} expect data from \code{\link{data4plot_Lollipop}}.
}
