% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RVP.R
\name{plot_RVP}
\alias{plot_RVP}
\title{Returns a ggplot2 object}
\usage{
plot_RVP(dat)
}
\arguments{
\item{dat}{A list including a dataframe, data, and labels, labs.}
}
\value{
a ggplot2 object.
}
\description{
\code{plot_RVP} expect data from \code{\link{data4plot_RVP}}.
}
