% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RVP.R
\name{plot_ts}
\alias{plot_ts}
\title{Returns a ggplot2 object.}
\usage{
plot_ts(dat)
}
\arguments{
\item{dat}{A list including a dataframe, data, and labels, labs.}
}
\value{
a ggplot2 object.
}
\description{
\code{plot_ts} expects data from \code{\link{data4plot_ts}}.
}
