% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waterfall.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plot waterfall chart.}
\usage{
plot_waterfall(dt, inputDriver)
}
\arguments{
\item{dt}{A dataframe with fields balance, DeltaValue, driver.}

\item{inputDriver}{A string.}
}
\value{
a ggplot object.
}
\description{
\code{plot_waterfall} get input from \code{\link{f_waterfall_input}}.
}
