% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treemap.R
\name{plotTree}
\alias{plotTree}
\title{Plot a treemap.}
\usage{
plotTree(dat)
}
\arguments{
\item{dat}{A dataframe with columns category, subcategory and total.}
}
\value{
a ggplot object.
}
\description{
\code{plotTree} takes input from \code{\link{dataTree}}.
}
