% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{repetitions_between_col}
\alias{repetitions_between_col}
\title{Check in hierarchy nested tables there is no rep of leveli within levelj.}
\usage{
repetitions_between_col(df)
}
\arguments{
\item{df}{a dataframe.}
}
\value{
the vector of the repeated values.
}
\description{
\code{repetions_between_col} not used.
}
