% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{rollmean}
\alias{rollmean}
\title{Compute the moving average of window x.}
\usage{
rollmean(x, window)
}
\arguments{
\item{x}{a numeric vector.}

\item{window}{integer, representing the width of the window.}
}
\value{
a numeric vector
}
\description{
\code{rollmean} take a vector as input returns the moving average vector.
The non computable data (the fist window-1 items) are replaced by NAs.
}
