% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{set_names_vect}
\alias{set_names_vect}
\title{Convert a dataframe representing a hierarchy into a 1D-list.}
\usage{
set_names_vect(df)
}
\arguments{
\item{df}{a dataframe representing a hierarchy.}
}
\value{
a named list to be used in a shiny::selectInput.
}
\description{
\code{set_names_vect} uses \code{\link{leadingDots}} and
\code{\link{flatten_in_order}}.
\code{set_names_vect} is used in UI to defined the list of top_suppliesvalue
dropdown box.
}
