% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{spread_to_list}
\alias{spread_to_list}
\title{Spread a tidyr nested dataframe to a named list.}
\usage{
spread_to_list(t)
}
\arguments{
\item{t}{a tidyr nested dataframe.}
}
\value{
a list
}
\description{
\code{spread_to_list} is used in \code{\link{multinest}}.
}
