% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppliesdash_input.R
\name{suppliesdash_data}
\alias{suppliesdash_data}
\title{Clean up app input data.}
\usage{
suppliesdash_data(myData)
}
\arguments{
\item{myData}{a list, containing 2 dataframe RVP and V.}
}
\value{
a list, myData with the 2 dataframe RVP and V amended.
}
\description{
\code{suppliesdash_data}
}
