% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppliesdash_input.R
\name{suppliesdash_rLevel}
\alias{suppliesdash_rLevel}
\title{function returning the rLevel, region hierarchy.}
\usage{
suppliesdash_rLevel(dat)
}
\arguments{
\item{myData}{a dataframe with columns RegionLeveli.}
}
\value{
a dataframe with only the columns RegionLeveli and the unique
  combinations.
}
\description{
\code{suppliesdash_rLevel}
}
