% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppliesdash_input.R
\name{suppliesdash_sLevel}
\alias{suppliesdash_sLevel}
\title{function returning the sLevel, supplies hierarchy.}
\usage{
suppliesdash_sLevel(myData)
}
\arguments{
\item{myData}{a dataframe with columns SuppliesLeveli.}
}
\value{
a dataframe with only the columns SuppliesLeveli and the unique
  combinations.
}
\description{
\code{suppliesdash_sLevel}
}
