% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{table2list}
\alias{table2list}
\title{Make a named list from nested dataframe.}
\usage{
table2list(ndf)
}
\arguments{
\item{ndf}{a nested dataframe (hierarchy).}
}
\value{
a nested list.
}
\description{
\code{table2list} creates a nested list from a nested dataframe ndf.
the nested data frame has at least 2 columns describing a tree (named
".*Leveli" colums), no other columns. From coarser to finer, from left
to right
}
\details{
Initially planned for use in parameter CHOICES in shiny::selectInput
but SHINY::inputSelect does not work as expected (the titles aren't clickable).
}
