% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ISenriched.R
\docType{data}
\name{test_ISenriched}
\alias{test_ISenriched}
\title{test dataset. ink sales and fit rate data.}
\format{a list of data frame
\describe{
\item{FR}{data frame fit rate}
\item{V}{data frame volume}
\item{RVP}{data frame Rev-Vol-Price}
}}
\source{
computation from inkSales and the hierarchy of region and product
}
\usage{
test_ISenriched
}
\description{
Ink sales formatted for Vol and revenue-volume-price, and fit rate.
}
\keyword{datasets}
