% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppliesdash_input.R
\name{unlist_unique}
\alias{unlist_unique}
\title{Was used for populating top_supplies value (suppliesInitValues).}
\usage{
unlist_unique(level1)
}
\arguments{
\item{level1}{a dataframe.}
}
\value{
a vector of strings with distinct items.
}
\description{
\code{unlist_unique} used to find the SuppliesInitValue from sLevel.
\code{unlist_unique} consume the result from suppliesdash_sLevel.
}
