% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS_common.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap text of a vector into a string of width "width".}
\usage{
wrap_text(x, width = 19)
}
\arguments{
\item{x}{a vector of strings.}

\item{width}{the maximum width of the string result}
}
\value{
a string of maximum width "width".
}
\description{
\code{wrap_text}
}
